Added subtle version tag
Designed to not be visible unless looking for it. The version is below the time on the Updated stack and is in a very dark grey `#222222`.

I realize this might be important if other people start to use the app, since I am going to want to change things as i use it and learn.